organise better config
