+++ backend/src/config/app.ts
@@ -1,0 +1,16 @@
+import express from 'express';
+import bodyParser from 'body-parser';
+import cors from 'cors';
+import * as dotenv from 'dotenv';
+import router from '../routes';
+
+dotenv.config();
+
+const app = express();
+
+app.use(cors());
+app.use(bodyParser.json({ limit: '20mb' }));
+app.use(bodyParser.urlencoded({ limit: '20mb', extended: true }));
+app.use(router);
+
+export default app;
