+++ backend/src/config/database.ts
@@ -1,10 +1,11 @@
-import mongoose from "mongoose";
+import mongoose from 'mongoose';
+import logger from '../logger';
 
 export default async () => {
-  try {
-    await mongoose.connect(`${process.env.DATABASE_URL}`);
-    console.log('MongoDB Connected..');
-  } catch (e) {
-    console.log(`Error connecting with MongoDB, ${e}`);
-  }
-}
+    try {
+        await mongoose.connect(`${process.env.DATABASE_URL}`);
+        logger.info('MongoDB Connected..');
+    } catch (e) {
+        logger.error(`Error connecting with MongoDB, ${e}`);
+    }
+};
