+++ backend/src/config/server.ts
@@ -1,28 +1,18 @@
-import express from 'express';
-import * as dotenv from 'dotenv';
-import bodyParser from 'body-parser';
-import cors from 'cors';
+import * as http from 'http';
+import app from './app';
 
-import router from '../routes';
+import logger from '../logger';
 
-const initializeServer = () => {
+const initializeServer = (): http.Server => {
     try {
-        const app = express();
-        dotenv.config();
-
-        app.use(cors());
-        app.use(bodyParser.json({ limit: '20mb' }));
-        app.use(bodyParser.urlencoded({ limit: '20mb', extended: true }));
-
-        app.use(router);
-
         const PORT = process.env.PORT || 5000;
 
-        const server = app.listen(PORT, () => console.log(`Server running on port: ${PORT}`));
-
-        return { app, server };
+        return app.listen(PORT, () =>
+            logger.info(`Server running on port: ${PORT}`),
+        );
     } catch (e) {
-        console.error(`Error launching server.. ${e}`);
+        logger.error(`Error launching server.. ${e}`);
+        return e;
     }
 };
